TTS_model/requirements.txt
nemo_toolkit[tts]
soundfile
torch

